"5:45pm.
```
createArray
	| actions size |
	actions := self actions.
	size := actions size.
	^ self createArray: size with: actions.
```

Now the actions here are nil. The reason why this is happening is because my grip on the code is too loose. It is no wonder it happened. Undefined variables plus lack of neat linear program flow makes it a nightmare to keep track of code. This kind of error literally never happened to me in either F# or Spiral.

Pharo IDE is nothing less than a miracle. Is there some reason no other language came along to break this monopoly?

I guess I will be finding this out.

5:55pm. I am completely losing my focus. Let me at least see if I can track this null bug down.

```
newAgent
	^ RPSAgent new actions: RPSAction subclasses
```

Ok, I get it. This thing creates an agent before I can assign it actions.

6:05pm.

```
initialize
	super initialize.
	regretSum := self createArray.
	strategySum := self createArray.
```

I can't do it like this. I am going to have to move the logic for `initialize` into `new:`. I am definitely going to make Pharo pay for making me go through this.

6:10pm.

```
new: actions
	x := self new.
	^self new
		regretSum: self createArray;
		strategySum: self createArray.
```

Oh, God this thing!

Now I do not have access to `createArray`.  This is absurd. Yeah, this is the standard OO experience. It is completely different than the one I have in Spiral.

6:15pm. Let me call it a day here before I get even more annoyed at this.

At this point, I am not even sure I know how to construct objects properly in Pharo. I should start with that tomorrow."