I want to talk about Pharo. This is what I wanted before I got distracted with that Git nonsense. I am writing this commit message in Iceberg, and it is restful being away from Visual Studio and my own journal. Out of sight, out of mind.
I think I have a good thing going now. It feels like for the past week I made very little real progress, instead most of my time was taken up doing things such as the ones today. So I did this performance comparison on a whim. Great. It cannot be called a complete waste of time.

But my goal was to do the algorithms in Pharo, and this goal I've been having a really hard time hitting.

A few days ago I got hit a programmer's block and lost my bearing, but I think it was only yesterday that I got good flow going. The day before that, even while I was depressed over losing most of my skills, I was thinking of how I could program better in Pharo and this effort inspired me. I got some good flow going yesterday.

I think if I am like that, I won't have any trouble surmounting the challenges ahead of me.

Right now I am thinking of the GVGAI port and the unpublished poker game with the GUI. Especially the poker game. It turned out to be a monster. It was what drove me to create Spiral. It was the thing that I threw in the towel on, knowing that I could not deal with the strenuous effort of interface maintenance for all my tooling.

I knew that I could deal with the ML library using Spiral, but I would need to deal with the interfaces past that. And now the time for that has come.

I knew that a regular level of effort would not be enough to deal with this.

Iit will take an absolutely fantastic effort. I will take all of my power and all of my skill and transform my programming. I will absorb everything that Pharo has to offer and reconnect with Spiral at the end of that.

...

I should be crying over this, but maybe the 6:45am wakeup and going straight into programming is a reflection of how I truly feel like this.

To me Spiral itself signifies freedom from the limitations of my previous language which was F#. In it I could do the ML library to a decent level of sophistication. But I have yet to break free of the limitations the platform itself imposes upon me. That platform is .NET.

Just imagine how much of a pain in the ass it would be if every time I wanted to print something, I had to mess with interrupt handlers for the Console? Right now that is the issue that I have with making GUIs in it. The only conclusion that I can take from that is that unless I take control of the system in its entirety, I will never manage to abstract this.

Pharo is possibly the best platform imaginable to get this level of control. Nothing will come close to its openness of design.

Though it might lack pattern matching, let bindings and lexical scope, I will be able to learn to like it. The language itself has a lot of principle. Like Spiral and the MLs, it is a crystallization of a style.

Right now I do not care about speed or static typing or anything else. I want to tap into the deep dynamism of Pharo. 

I want to see what I am capable of when I am not held back by restraints on expressiveness or on my imagination. I want to get more of that flow going.

Think of it as finishing a game and starting a whole new one. By abandoning the things that I've done, I make room for the things that are to come. I want to leave behind my meaningless responsibilities and the goals that I cannot possibly reach. I just want to keep on growing. I want my goals to make me better. I want the power that lies at the end of all of this.