It is time do the visualizations. I really slept well tonight and feeling ready to start the day.
Today's mission is Roassal. I will finally use it for real. All the examples in the book are pretty elaborate and it is not clear to me how to make a simple stacked bar chart.

...Let me just start this thing already. I've been chilling for over an hour just thinking. Let me crank up the heat.

The past four years have just been warm up for this moment. The practice I got on Spiral will serve me well when the time comes to take apart Pharo. I need to go down this path.

There is no rushing this. Taking apart Pharo is not something that I will be able to just do anyway. It will be a slow and steady accumulation of experience. Best to take it easy and enjoy myself as I go along. If I just push myself I'll burn out and that would be much worse than taking extra time to get to my goal.