Trying to start at part two directly in the Builders is rough. It is a bit too much.
http://agilevisualization.com/AgileVisualization/PharoIntroduction/0102-PharoIntroduction.html

Instead this is where I need to begin. It is a tutorial on Pharo, but also covers the very basics of Roassal. I need to start off with views...

And I hear thunder right now. It has been a while since I had to turn off. For the safety of my rig, let me do so.