"7:10am. Right now I am still chilling. Got up 10m ago. Let me have some fun and then I will start.
7:15am. https://www.reddit.com/r/MachineLearning/comments/azvbmn/n_openai_lp/

Honestly, I am envious of OpenAI. I wish that I could make money using media hype rather than having a restriction on having to rely on raw skill. With this they will move further and further away from the task of mastering the fundamentals that they need to actually make AGI.

IBM's Deep Blue might have beaten Kasparov, but it is not they who made chess engines ubiquitous. It is the refinement that they need to get to their goal.

If it turns out that having the biggest computer is what they need then they will win the race. If it turns out that what they need is programming skill then their lack of patience will be their downfall. It needs to be remembered that an atom bomb is a different thing than a big enough of pile of TNT, even though both are in the business of producing explosions.

Nobody has it easy. Especially me.

All I wanted is to master poker and at some point my focus became languages. Now I am pretty much going to be working on an OS.

7:30am. I am amazed at my continual ability to generate flow. I guess having the right goal really makes a difference. Let me finish chilling and then I will watch the Glamorous Toolkit talk."