"7:45am. Let me go for the talk. Damn it is freezing here. I really wasted the morning, but that is how it should be. Yesterday I was not as tired as I would have been after a full day of programming.
7:50am. https://youtu.be/qeYEais1eT4?t=1130

"This is something that Smalltalk should have excelled, and that is data science."

Yes, yes...that is how things should have went. Either way, I can't really see any of the MLs doing it. Spiral is really suited mostly for the low level stuff. It is Smalltalk that should have won in this field. Smalltalk's vision is just so different.

9am. http://elbenshira.com/blog/the-end-of-dynamic-languages/

I've really been obsessed with OpenAI threads. I've wasted enough of that. Let me finish reading this and then I will start programming. I will figure out how to position the System Browser properly.

9:35am.

```
openInWindowLabeled: aString

	^self openInWindowLabeled: aString inWorld: self currentWorld
```

Ok, it is here that the thing gets moved.

9:45am.

```
translatedAndSquishedToBeWithin: aRectangle
	"Return an adjustment of the receiver that fits within aRectangle by
		- translating it to be within aRectangle if necessary, then
		- reducing its size, if necessary"

	^ (self translatedToBeWithin: aRectangle) squishedWithin: aRectangle
```

Hmmm, so this is the method that squishes the window.

10:35am. I've had to take a short break. Stop with those dumb UBI threads.

Personally, I do not see why UBI could not be used to replace the tax system. Tax avoidance using loopholes is a problem when it comes to taxing large corporations. So it is keeping track of hugely complicated tax systems. Letting inflation take care of things might be an alternative to explicit taxation. There is the risk of inflation spiraling out of control, but I am sure that would never happen. *grin*

More seriously, that sort of thing is always a risk with fiat currency. Gold is really no exception either.

So UBI is hardly unique in that regard, the biggest impact would be that a lot of tax lawyers would need to find something else to do. Maybe become economists?

/pol/ exists to push new political ground, so maybe this was inevitable. Maybe they should take up the mantle of spearheading UBI.

10:50am. Ok, focus me. I've ascertained something.

Let me go to `openInWindow`.

```
openInWindowLabeled: aString inWorld: aWorld
	"Changed to include the inset margin in the bound calculation."

	| window extent |
	window := (SystemWindow labelled: aString) model: nil.
	window 
		" guess at initial extent"
		bounds:  (RealEstateAgent initialFrameFor: window initialExtent: self fullBounds extent world: aWorld);
		addMorph: self frame: (0@0 extent: 1@1);
		updatePaneColors.
	" calculate extent after adding in case any size related attributes were changed.  Use
	fullBounds in order to trigger re-layout of layout morphs"
	extent := self fullBounds extent + 
			(window borderWidth@window labelHeight) + window borderWidth +
			(window class borderWidth * 2 @ (window class borderWidth + 1)). "include inset margin"
	window extent: extent.
	aWorld addMorph: window.
	window activate.
	aWorld startSteppingSubmorphsOf: window.
	
	window announceOpened.
	^window
```

This is where all the repositioning happens.

`bounds:  (RealEstateAgent initialFrameFor: window initialExtent: self fullBounds extent world: aWorld);`

This part actually does squishing of the windows, but 

```
	extent := self fullBounds extent + 
			(window borderWidth@window labelHeight) + window borderWidth +
			(window class borderWidth * 2 @ (window class borderWidth + 1)). "include inset margin"
```

By the time this gets to here, that thing is lost.

11am. I am convinced that this sort of thing might be a bug. That `self fullBounds extent` should really be `window fullBounds extent`. The intent behind the code does not make sense.

11:05am. Eh, I crashed the image again.

11:10am. Ok, once I do the change I suggested, the resizing is almost correct. For some reason it is not taking the translucent bottom tab into account when sizing windows. I think it would be better if it did.

11:15am. Let me step through the `RealEstateAgent` so I can determine how it actually figures out the rectangle.

...I crashed the image again. I should not have put `halt` into the `initialFrameFor:initialExtent:world:` method.

Let me try again.

```
cascadeFor: aView initialExtent: initialExtent world: aWorld

	| position allowedArea |
	allowedArea := self maximumUsableAreaInWorld: aWorld.
	position := aWorld currentWindow isMorph 
		ifFalse: [ aWorld center - (initialExtent / 2)]
		ifTrue: [ aWorld currentWindow position + 20].
	^ (position extent: initialExtent)
		translatedAndSquishedToBeWithin: allowedArea
```

This is the method I was looking for.

And the allowed area is `allowedArea := self maximumUsableAreaInWorld: aWorld.` This. Let me debug into this.

```
maximumUsableAreaInWorld: aWorldOrNil

	| allowedArea |
	
	allowedArea := 0@0 extent: self standardSize. "fallback default size"
	
	aWorldOrNil ifNotNil: [
		allowedArea := aWorldOrNil displayArea intersect: aWorldOrNil visibleClearArea ifNone: [ allowedArea ]].
	^allowedArea
```

`aWorldOrNil displayArea intersect: aWorldOrNil visibleClearArea ifNone: [ allowedArea ]]`

This is the part that I am looking for.

11:25am.

```
boundingBox
	clippingBox == nil
		ifTrue: [clippingBox := super boundingBox].
	^ clippingBox
```

I've seen this while debugging. Old code is used here. I should just replace this with `ifNil:`

11:30am.

```
clearArea
	"Answer the clear area of the receiver. It means the area free  
	of docking bars."
	| visTop visBottom visLeft visRight |

	visTop := self top.
	visBottom := self bottom.
	visLeft := self left.
	visRight := self right.

	self dockingBars
		do: [:each | 
			(each isAdheringToTop and: [each bottom > visTop])
				ifTrue: [visTop := each bottom].

			(each isAdheringToBottom and: [each top < visBottom])
				ifTrue: [visBottom := each top].

			(each isAdheringToLeft and: [each right > visLeft])
				ifTrue: [visLeft := each right].

			(each isAdheringToRight and: [each left < visRight])
				ifTrue: [visRight := each left]
		].

	^ Rectangle
		left: visLeft
		right: visRight
		top: visTop
		bottom: visBottom
```

```
dockingBars
	"Answer the receiver's dockingBars"
	^ self submorphs
		select: [:each | each isDockingBar]
```

I am betting it is not counting that translucent tab as a docking bar.

Let me take a break. I am into it, but it seems that investigating this thoroughly will take me a while.

11:50am. Forgot to commit."