I finally got up a little later for once. And I have nothing to do but program. No need for replies and no distractions.
This is the way it should be.

Going forward things will be a different compared to the past. From 2015 to 2019,  I had a goal which was to master the programming aspects and make a library, then just go into poker. In a sane world, all the tools that I'd need would have already been provided to me in 2015. I wanted a straightforward path to my desire, but somehow I ended up working on a language and a library for a very long time.

Early 2016 was when all the pieces should have been in place for my plan to work. Instead it is 2019 and I still haven't gotten the things that I need. Are simple, game-like plans impossible in the real world? That is what a part of me is wondering. My life is a wreck because of it, but it is fine. I accounted that it might turn out this way. It is the result of my sin. That sin being the lack of power.

Ultimately, I have no choice but to switch to Pharo if I want to arrest the trend of having all my tooling be done from the command line. This will cost me significantly in time. But the programs I made in 2016 cast a long shadow. I cannot escape from the need to do everything perfectly if I want to win.

OpenAI might be able to throw a media circus and ride the hype wave, but for individuals only skill will do.

I guess for being a good programmer, every single aspect of it needs to be mastered. Any hole will get punished eventually. Trying to make a ML library in F# really punished me. So did trying to make graphical interfaces. Or that GVGAI port that really challenged the extensibility aspects of it.

At this point I am just about ready to make the move to a dynamic language and the only one that will suffice will be a Smalltalk variant like Pharo. Eventually, the 10x performance gap between Spiral and Pharo will start to squeeze, but at the very least even though it will be time consuming to bring in Spiral, I cannot say that I do not have a solution for that. Pharo is just so perfect. It is so moldable and it feels like I can do anything in it. Every single aspect of it is at my disposal and I have a deep appreciation for that.

I am also happy that the gap is only 10x - Pharo is quite fast as a language.

All the deadlines I have set for myself have long been crossed. But I knew that I was the sort of person to let this happen based on my trading experience. This time, though I failed at least I will retain the skills that I made for my effort. Unlike trading which will die as soon as I crack credit assignment, programming is dealing with fundamental aspects of reality. This power that is useless at the moment will never fail me. Programming is eternal, but trading died for me the moment I imagined the self improvement loop from the inside.

Unlike before, I do not have a concrete goal like making a ML library or implementing a fast RNN. I did those things and it got me nowhere.

Right now my goal is not even poker so much - I will simply aim to get better at programming and that will be it. I will go in the direction of making interfaces and a poker agent for them, and the plan will be to just let my accumulate experience and effort make their impact. If this approach is useless, I guess the only conclussion will be that networking is the end all of life and deep pocketed companies like DeepMind and OpenAI will be the ones to cause the Singularity rather than individuals such as myself.

I need to get better.

When one avenue gets exhausted, I will just pick another. That is how I will reach my goal. There is only a limited amount of avenues left, but it will be a long journey nonetheless.

Staged functional programming as spearheaded by Spiral is dead. But it will make a revival at some point. It will be through me and my own effort.

After 4 years of this, I think I have all the basics down. What I need to attain is deep integration of all my tooling from here on out. That power once attained will be unconquerable and absolute.

Absolute will to reach the absolute power.

There is no need to waver. Today I will make just a small step towards that.

Let me resume the Roassal tutorial.