"1:15pm. Let me resume, I wasted enough time. It is time to get into the swing of things.
1:25pm. Hmmm, I thought to make `NormalizedArray` inherit from `IntegerArray` instead of the standard one, but it is not working. Well, it is another thing to investigate.

2:10pm. Stop reading the F# sub.

http://agilevisualization.com/AgileVisualization/Roassal/0104-Roassal.html

It is amazing that I keep getting distracted given that I am into this to an extent.

2:25pm.

```
	v := RTView new.
	es := (RTBox new color: Color blue trans) 
			elementsOn: (1 to: 20).
	v addAll: es.

	RTNest new for: es add: [ :group :model |
		group addAll: ((RTBox new color: Color red) elementsOn: (1 to: model)).
		RTGridLayout on: group ].

	RTGridLayout on: es.
	v
```

Ok, I am a bit bored of this, but I feel that it is sinking in for me. I get this.

3pm. 'Dragging the left box forward the'.

Spelling errors. `forward` should have an `s`.

https://www.reddit.com/r/MachineLearning/comments/b0rdsi/d_irresponsible_anthropomorphism_is_killing_ai/
https://www.reddit.com/r/MachineLearning/comments/b0uvv2/d_the_economist_article_about_deepminds_culture/

I'll check these threads out later. Let me continue with the chapter.

3:25pm. `Normalization is essential for rendering curve and graph. Consider the script:`

Spelling errors. Should be `curves and graphs`.

3:45pm. 

```
	v := RTView new.
	es := (RTBox new size: 15; color: Color blue) elementsOn: #(4 5 1 2 3 5).
	v addAll: es.

	es @ RTPopup.

	RTMetricNormalizer new
		elements: es;
		alphaColor: 0.3;
		normalizeHeight: #yourself min: 0 max: 80.

	RTHorizontalLineLayout new alignBottom; on: es.

	es @ RTLabeled.

	v
```

With this example I have enough information on how to implement my own bar charts. I already know how to stack boxes vertically. With this I will be able to visualize the dataflow for simple RL algorithms. NNs with millions variables will be a lot harder though.

4:05pm.

```
	values := #(35 26 5 18 17 60 36 13 16 28 29 3 8 0 3 2 14 
				12 13 17 1 9 3 3 4 1 1 1 2 1 1 61).
	v := RTView new.
	n := RTMultiLinearColorForIdentity new objects: values.
	shape := RTBox new color: n; size: #yourself.
	es := shape elementsOn: values.
	v addAll: es.
	RTHorizontalLineLayout new gapSize: 1; on: es.
	RTAlignment new elements: es; bottom.
	v
```

Really, it is amazing what Roassal can do. I will master all of its aspects thoroughly.

5pm. Finally done with chapter 4 of the Roassal book. I think I absorbed it this time. That last chapter was cool. Roassal is definitely a vast improvement on FsLab.

The particular chapter in the book that I just went through was quite large. Let me go through one more and I will call it a day."