"10:05pm. Let me finally start. I chilled way too much today. Let me go through the part one before breakfast. That will leave the builder chapters for later.
10:25pm.

```
	v := RTView new.
	
	shape := RTCompositeShape new.
	shape add: RTLabel new.
	shape add: (RTBox new
		color: Color transparent; borderColor: Color black).
	es := RTGroup
		with: (shape elementOn: RTFlowLayout)
		with: (shape elementOn: RTHorizontalFlowLayout)
		with: (shape elementOn: RTGridLayout)
		with: (shape elementOn: RTCellLayout).
	v addAll: es.
	
	RTNest new
		for: es
		inShape: #second
		add: [ :group :layout |
			group addAll: ((RTBox new size: [:m | m * 10])
				elementsOn: (1 to: 6)).
			layout new on: group.
		].
	
	RTCellLayout new gapSize: 10; on: es.
	
	(v elements allButFirst: 4) @ (RTLabeled new
		color: Color black; center).
	v
```

This is such a complex example. The labels inside the boxes do not show up properly. I'll leave debugging this for later.

10:50am.

http://brrt-to-the-future.blogspot.com/
Stuff about VM and JIT internals

I'll keep this blog in mind.

11am.

```
	v @ RTDraggableView @ RTZoomableView.
```

Oh, if you make the view zoomable, the mouse wheel can be used to zoom and not just the buttons.

11:10am.

```
elementFromModel: anObject
	"Return an element that corresponds to the object provided as parameter. If none is found, then return nil"
	
	| t key|
	self elementFromModelCache at: anObject ifPresent: [ :v |
		v isCollection ifFalse: [ ^ v ] ].

	"We have this line because of a bug of the implementation of CompiledMethod>>#="
 	t := (anObject class == CompiledMethod) 
				ifTrue: [ self detect: [ :el | el model == anObject ] ifNone: [ nil ] ]
				ifFalse: [ self detect: [ :el | el model = anObject ] ifNone: [ nil ] ].
	"We have this hack for nil objects because in visualWorks a dictionary cannot have nil as a key"
	key := anObject isNil ifTrue: [ 'KeyForNilObjects' ] ifFalse: [ anObject ].
	self elementFromModelCache at: key put: t.
	^ t
```

You can find interesting things if you look at the source.

Let me try removing that hack.

Yeah, it works. Presumably the bug has been fixed.

11:15am.

```
	classes := RTLayout withAllSubclasses.
	b := RTMondrian new.
	b shape ellipse color: Color purple; size: 5.
	b nodes: classes.
	b edges connectFrom: #superclass.
	b layout sugiyama.
	b
```

This thing is broken. It does not minimize edge crossings at all.

11:20am.

```
v := RTView new.
es := 
	RTEllipse new 
		color: Color purple;  
		elementsOn: RTLayout withAllSubclasses.
		
eb :=
	RTEdgeBuilder new
		view: v;
		connectFrom: #superclass.

v
```

Ah, I will play with this later. There are so many ways to do the same thing in Roassal.

Let me finish the chapter so I can go have breakfast.

11:30am.

```
	b := RTNameCloud new 
		addString: 'open
		| v shape |
		v := RTView new.
		shape := RTLabel new height: [ :assoc | assoc value ]; text: #key.
		v addAll: (shape elementsOn: table associations).
		RTFlowLayout on: v elements.
		v open'.
	b
```

This is an interesting example. I am betting there is a way to use the name cloud in a regular fashion. But this is a good demo of Pharo's dynamic code execution capabilities.

```
	b := RTMondrian new.
	b nodes: RTShape withAllSubclasses.
	b shape line color: Color blue trans.
	b edges 
		connectToAll: [ :c | c dependentClasses copyWithout: c superclass ].
	b layout 
		force.
	b
```

I kind of like how Mondrian works. It is a high level interface to the stuff I've been seeing so far.

11:45am. http://agilevisualization.com/AgileVisualization/Interaction/0107-Interaction.html

This is taking me longer than I'd thought. Let me stop here, I'll finish part one after the break."