"2:10pm. http://agilevisualization.com/AgileVisualization/Interaction/0107-Interaction.html
Fuuu, I am finally done with this chapter.

The next comes part two. I am getting the feeling for this, but this kind of learning is just such a miserable way of doing it.

http://agilevisualization.com/AgileVisualization/Builder/0201-Builder.html

Let me move on to builders.

Oh, there is a whole chapter dedicated to charting. Great.

2:15pm. Focus me, focus.

`Frequently, visualizations have the same limitation than software artifact`

Those last 3 words are in need of a rewrite.

`The System Browser is open from the World menu.`

Should be `is opened`."