"2:40pm.
```
	renderIn: aView	
		aView add: (RTLabel elementsOn: ' ').
		objects 
			do: [ :object | aView add: (RTLabel elementOn: object) ].
	
		metrics do: [ :assoc |
			aView add: (RTLabel elementOn: assoc key).
			objects do: [ :object |
				| value |
				value := assoc value rtValue: object.
				aView add: ((RTEllipse new size: value) element) ] ].		
			
		RTCellLayout new 
			lineItemsCount: objects size + 1;
			on: aView elements
```

I've been really focusing on this part. Ok, I think I understand it.

The key to understanding big pieces is to really understand small pieces. The understanding of the small is a prerequisite to understanding how the large comes together.

3pm. Ok, focus me. Just one more example and I am done with the chapter. I will take a break after that.

I am starting to get a sense not just of Roassal, but how Pharo works. I'll accept the strange and the bad, and take it along with the good. These ways of constructions builders are quite expressive.

3:10pm. http://agilevisualization.com/AgileVisualization/Mondrian/0202-Mondrian.html

Finally done with the chapter. Uah.

I want a break as this is pretty tedious, but let me go on for a while longer."