"8:15am. Focus me. I've gone through one example, but I am getting lost in my imagination.
8:30am. https://johncarlosbaez.wordpress.com/2019/03/15/algebraic-geometry/

I'll consider this later. Let me get back to the examples.

8:45am.

```
	b := RTGrapher new.

	d := RTMultipleData new.
	d barShape color: Color blue.
	d points: #( #('hello' 1 2 1) #('world' 2 4 2) #('bonjour' 3 5 4) #('Gutten Morgen' -1 4 -5)).
	d addMetric: #second.
	d addMetric: #third.
	d addMetric: #fourth.
	
	"Rotated text"
	d barChartWithBarTitle: #first rotation: -30.

	b add: d.

	b
```

I am near the end. Now if only one of the following examples would show how to stack the data points that would be great.

8:55am.

```
	methods := Collection withAllSubclasses flatCollect: #methods.
	trachelMethods := TRObject withAllSubclasses flatCollect: #methods.

	b := RTGrapher new.

	"Data set 1"
	ds1 := RTData new.
	ds1 label: 'collection methods'.
	ds1 dotShape circle color: Color red trans.
	ds1 points: methods.
	ds1 x: #numberOfLinesOfCode.
	ds1 y: [ :m | m sourceCode size ].
	b add: ds1.

	"Data set 2"
	ds2 := RTData new.
	ds2 label: 'trachel methods'.
	ds2 dotShape circle color: Color blue trans.
	ds2 points: trachelMethods.
	ds2 x: #numberOfLinesOfCode.
	ds2 y: [ :m | m sourceCode size ].
	b add: ds2.

	b axisX title: 'Number of lines of code'; noDecimal.
	b axisY title: 'Number of characters'; noDecimal.

	"We build the legend, and indicate it has to be located below"
	b legend below.

	b
```

Had to change `getSource` to `sourceCode` here as well.

One of the previous examples used dates so it did not work.

9am. Oh crap, this chapter cuts off before the end.

Mailed the guy to tell him that. No way around it, let me go through the last two chapters and then I will be almost done with the book. Only part 3 will be left after that.

I'll finish this thing today.

9:10am. http://agilevisualization.com/AgileVisualization/LegendBuilder/0204-LegendBuilder.html

Hmmm, this is a short chapter as well. Only a single example.

http://agilevisualization.com/AgileVisualization/BuilderComposition/0205-BuilderComposition.html

Only this one is left.

9:30am.

```
	data := #(20 30 50 100).

	c := RTComposer new.

	"First visualization"
	g := RTGrapher new.
	g extent: 200 @ 200.
	g view: c view.
	ds := RTData new.
	ds interaction popup.
	ds barShape color: Color blue.
	ds points: data.
	g add: ds.
	g axisX noTick; noLabel.
	g build.
	c group: #graph.

	"Second visualization"
	b := RTPieBuilder new.
	b view: c view.
	b interaction popup.
	b objects: data.
	b slice: #yourself.	
	b labeled.
	b build.
	c group: #pie.

	"Layouting"
	c move: #graph onTheLeftOf: #pie.
	c nameGroup: #graph as: 'Values as a bar chart'.
	c nameGroup: #pie as: 'Values as pie areas'.
	
	RTMetricNormalizer new
		view: c view;
		objects: data;
		distinctColor.

	c view
```

I just realized, but `interaction popup` does not work for this example.

9:40am. Focus me. I only have one example left in part two. Then comes part three. I think I will just skim that one.

9:45am. Ok, enough of that example.

10am. The examples are broken in Pharo 7 it seems.

I guess if I want everything to work, I should get Roassal with its image.

http://agilevisualization.com/AgileVisualization/OpenStreetMap/0304-OpenStreetMap.html

Right now I am just skimming these as I said I would. I am going to have to ask on how to do stacked bar charts.

10:30am. Done with the examples. I might as well stop here for a bit.

The mission when it is time to resume will be to figure out how to make stacked barcharts. There should probably be a way to do it somewhere in the `RTGrapher`. I'll have to poke around the class a bit to figure it out."