"12:30pm. Let me finally resume. My rest is so restless. I just wonder the chans aimlessly looking for the fix that never comes.
My despair is high. Let's do this.

I'll fiddle with the grapher in order to figure out how to make stacked bar charts.

12:30pm. http://agilevisualization.com/AgileVisualization/Grapher/0203-Grapher.html

The most sensible place to start would be in several metrics under one data point example.

12:40pm. 

```
thisIsAnObsoleteClass
	"please, use RTHorizontalMultipleData instead"
```

I had no idea the class was obsolete. But this comment is quite fortunate for me. If there is a horizontal multiple data, there is probably a vertical or a stacked one.

12:55pm.

```
	b := RTGrapher new.

	d := RTVerticalMultipleData new.
	d barShape color: Color blue.
	d points: #( #('hello' 1 2 1) #('world' 2 4 2) #('bonjour' 3 5 4) #('Gutten Morgen' 1 4 5)).
	d addMetric: #second.
	d addMetric: #third.
	d addMetric: #fourth.
	
	"Rotated text"
	d barChartWithBarTitle: #first rotation: -30.

	b add: d.

	b
```

Ah, wonderful. This automatically gives me the stacked bar chart.

But damn do the internals of Roassal feel difficult. All the complexity is in essentially building the AST for the visualizations and the entire library is significantly hamstrung by the fast the language has no support for immutability.

How many days of my life did I spend on this?

I think I started Pharo like two weeks ago. Just where did this time go?

Let me go through the journal for a few minutes. I am curious.

I need a mini-break."