"3:45pm. Focus me.
```
main := GraphingRunner new: (BlottoAction numFields: 2 numSoldiers: 1).
```

Let me back this up here.

3:50pm. No just adding things is not enough. I need to figure out how to signal the update.

Hmmm...Let me go through the init with the debugger.

This is the issue with these DSLs. They tend to be restricted and need modification.

3:55pm.

```
addedInBuilder: aBuilder
	super addedInBuilder: aBuilder.
	aBuilder axisX noLabel; noTick.
	
	elements := RTGroup new.
	points do: [ :p | 
		| values es |
		values := metrics collect: [ :m | m rtValue: p ].
		es := barBuilder elementsOn: values.
		es doWithIndex: [ :e :index | e trachelShape color: (self colorsToUse at: index) ].
		groupedElements add: es.
		elements addAll: es
		 ] separatedBy: [ elements add: ((RTBox new size: 0) elementOn: 0) ].
	self addAll: elements
```

The grapher method add calls this. This is what compiles the points to trachel shapes.

4:05pm. I am going to abandon how I intended to do this. I can make this work, but since every element needs to be a chart, I'll batch them in number of 100 or something like that. That would be the best way to do this.

Otherwise I will have to make an entirely different class just to I can get smooth charting. No thanks to that.

That having said, the way this gets the metrics...it will struggle mightily with what I had in mind.

Let me try a few things.

Agh, what a pain this is. But I have to do this. I must.

4:15pm. Adding multiple elements in `RTGrapher` does not work.

4:20pm. Ah, I see that the text at the bottom is smushed. They might be overlaid. Maybe I should apply a layout to the grapher?

```
	RTCellLayout on: b view.
```

Nah, neither this nor just `b` works.

How do multiple elems work here again?

4:25pm. I give up. I think I will just ask on the SO.

4:45pm. https://stackoverflow.com/questions/55198661/how-do-i-dynamically-add-elements-to-the-roassal-rtgrapher-instance

```
    Object subclass: #MultiData
    	instanceVariableNames: 'b'
    	classVariableNames: ''
    	package: 'CFR-Extensions'

    initialize
    	b := RTGrapher new.
    
    	b add: (self makeD: #('hello' 1 2 1)).
    	b add: (self makeD: #('test' 1 2 11)).
    
    	b

    makeD: first
    	| d |
    	d := RTVerticalMultipleData new.
    	d barShape color: Color blue.
    	points := OrderedCollection new.
    	points add: first.
    	d points: points.
    	d addMetric: #second.
    	d addMetric: #third.
    	d addMetric: #fourth.
    	
    	"Rotated text"
    	d barChartWithBarTitle: #first rotation: -30.
    	^d

The above is essentially the `Several metrics per data point` [example](http://agilevisualization.com/AgileVisualization/Grapher/0203-Grapher.html) from the Roassal book factored into two methods. Rather than just visualizing a static dataset I've been looking into ways of trying to add data as the program runs. I want to visualize the trace of the parameters for a tabular RL agent.

What happens when I display the graph in the inspector is that only the latest element shows up as a chart. There is some overlaying in the labels though that should not be there.

Originally I wanted to do something like pass an `OrderedCollection` of points, but the way `RTVerticalMultipleData` compiles them into Trachel elements makes such a scheme invalid, so I've thought to batch the data instead before adding it as an element.

The fact that the above does not work strikes me as a bug. Apart from fixing this, I am wondering if there is a better way to visualize dynamic data?
```

Hmmm...Let me post this and then I'll take a break.

4:55pm. Finally posted it. Yeah, I could debug this, but I do not feel like it right now.

Before I stop for the day, I want to see what NN & GA support in Roassal is about.

5:05pm. It seems they are not NN or GA specific visualizations like I'd hoped for but rather a crappy ML library that the Roassal author did at some point. Useless.

5:10pm. To be honest, I do not feel like messing with Roassal anymore.

Literally everything in Pharo is beta quality at best. I suppose I could try debugging this myself tomorrow.

5:15pm. But rather than do that, I am interested in investigating Raptor's Studio.

The plan was really to continue with CFR, but forget that. I do not want to even think about doing that in Pharo at the moment. Maybe I will go back to F# by the end of the month and try doing an UI using the Elm pattern there.

Now that I am really focused on UIs, the Elm pattern is a real possibility. In 2016 I did not want to bother implementing it and felt that it is something that should be handed to me, but now I am more open to changes.

5:20pm. I have no choice but to move forward with my studies rather than projects since things are going like this.

At any rate, it does not seem like Roassal will be that useful to me.

I want to visualize large quantities of dynamic data, while it is focused on small quantities of static data.

5:25pm. Pharo definitely does not feel like it has been in the works for over 10 years. It feels more like 2 years at best. A dynamic language like it does not support the rapid refactoring that the coredevs want to take up.

I am going to hae to study the platform, but I am still not convinced that I want to use it.

Let me commit here. I think I will call it a day."