"Let me install GT.
...Wow, I am impressed. It feels slow as molases and I managed to crash it in like 5 seconds. I literally just typed a few things into the opening window and that is what happened. Let me restart the image.

...Let me increase the resolution to max. I can't deal with Pharo like this.

9:45am. It seems GT has a bug where pressing backspace causes it to become toggled. The thing looks nice, but the whole system feels Emacs sluggish.

I can't use this. This is super beta quality. Just how could it possibly have a bug like this? Just what are the devs doing?

So far Pharo is really challenging my determination that Singularity will not be be caused from the command line.

9:55am. What is my plan for now? Let me see if I can resolve that Roassal bug. Today I want to spend some time debugging it. It will be good exercise for me.

Though I am sure it was not even close to its intention Pharo is really making me see the benefits of static typing. As a language it could really benefit from it."