"10:15am. Ugh...Well, I think I know. But doing this kind of debugging is rapidly draining my will to live. I really understand why Haskell zealots are so hung up on purity after this.
Let me give it a show again. I am so tired of this I actually closed the image.

10:15am.

```
addAll: someElements
	self view addAll: someElements
```

Let me apply the vertical layout to this.

```
addAll: someElements
	| view |
	view := self view.
	view addAll: someElements.
	RTCellLayout view.
```

Let me try it like this.

```
addAll: someElements
	| view |
	view := self view.
	view addAll: someElements.
	RTCellLayout on: view.
```

No wait, I think the layout is suppose to be applied on the elements.

```
addAll: someElements
	| view |
	view := self view.
	view addAll: someElements.
	RTCellLayout on: view elements.
```

Maybe like this?

10:20am. No, I give up. Pharo is going to kill me at this rate.

There is just too much clashing between the style I want to program in and the style it provides. I want its IDE experience, but I also want a language that takes into account the last 40 years of language design improvements.

Let me go into academic mode. I'll read the old Smalltalk book on its implementation. They I will do some studying of various runtimes and such.

It is clear to me that CFR is not going to get done at this rate.

The thing is, I could debug this if I put enough effort into it, but the way it is structured pretty much ensures that I won't be able to figure out how Roassal works unless I figure out the whole system. Pervasive mutability manifests as pervasive coupling.

I have such a finely tuned style of programming in both Spiral and F#, but Pharo pretty much demands of me that I flush it down the drain. The simplicity of Pharo ensures that I cannot see the cues that give me insight into the greater structure.

At this point, maybe I should consider one of the web languages and port Spiral to web assembly?

10:30am. Let me make this the last commit for a while in the Pharo repo. I'll go back to blogging in the Spiral repo."